move app setup to setup.js
